Refactor BlogGrid component
